all green test

+++ db/migrate/20201003223537_create_payments.rb
@@ -2,7 +2,7 @@ class CreatePayments < ActiveRecord::Migration[6.0]
   def change
     create_table :payments do |t|
       t.integer :status
-      t.float :value, default: 0, null: false
+      t.integer :value, default: 0, null: false
       t.string :token
       t.string :purchase_details_url, default: ''
       t.string :payment_url, default: ''
