+++ db/schema.rb
@@ -17,7 +17,7 @@ ActiveRecord::Schema.define(version: 2020_10_03_223537) do
 
   create_table "payments", force: :cascade do |t|
     t.integer "status"
-    t.float "value", default: 0.0, null: false
+    t.integer "value", default: 0, null: false
     t.string "token"
     t.string "purchase_details_url", default: ""
     t.string "payment_url", default: ""
